+++ src/main/java/com/ditrit/letomodelizerapi/service/UserServiceImpl.java
@@ -5,11 +5,16 @@ import com.ditrit.letomodelizerapi.model.error.ErrorType;
 import com.ditrit.letomodelizerapi.model.user.UserRecord;
 import com.ditrit.letomodelizerapi.persistence.model.User;
 import com.ditrit.letomodelizerapi.persistence.repository.UserRepository;
+import com.ditrit.letomodelizerapi.persistence.specification.SpecificationHelper;
 import jakarta.servlet.http.HttpSession;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
@@ -18,6 +23,7 @@ import java.net.URISyntaxException;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
+import java.util.Map;
 
 /**
  * Implementation of the UserService interface.
@@ -31,6 +37,12 @@ import java.net.http.HttpResponse;
 @Transactional
 @AllArgsConstructor(onConstructor = @__(@Autowired))
 public class UserServiceImpl implements UserService {
+
+    /**
+     * The attribute key used for retrieving the user's login information from a session or context.
+     * This constant defines how the login attribute is referenced within the application.
+     */
+    private static final String LOGIN_ATTRIBUTE = "login";
 
     /**
      * The UserRepository instance is injected by Spring's dependency injection mechanism.
@@ -54,11 +66,11 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User getFromSession(final HttpSession session) {
-        String login = session.getAttribute("login").toString();
+        String login = session.getAttribute(LOGIN_ATTRIBUTE).toString();
 
         return userRepository
             .findByLogin(login)
-            .orElseThrow(() -> new ApiException(ErrorType.AUTHORIZATION_ERROR, "login", login));
+            .orElseThrow(() -> new ApiException(ErrorType.AUTHORIZATION_ERROR, LOGIN_ATTRIBUTE, login));
     }
 
     @Override
@@ -80,5 +92,19 @@ public class UserServiceImpl implements UserService {
             Thread.currentThread().interrupt();
             throw new ApiException(e, ErrorType.INTERNAL_ERROR, "picture", user.getPicture());
         }
+    }
+
+    @Override
+    public Page<User> findAll(final Map<String, String> filters, final Pageable pageable) {
+        return this.userRepository.findAll(new SpecificationHelper<>(User.class, filters), PageRequest.of(
+                pageable.getPageNumber(),
+                pageable.getPageSize(),
+                pageable.getSortOr(Sort.by(Sort.Direction.ASC, "name"))));
+    }
+
+    @Override
+    public User findByLogin(final String login) {
+        return this.userRepository.findByLogin(login)
+                .orElseThrow(() -> new ApiException(ErrorType.ENTITY_NOT_FOUND, LOGIN_ATTRIBUTE, login));
     }
 }